+++ MenuElements/average_class_container.lua
@@ -144,7 +144,7 @@ function average_class_container.updateMenuElement(scroll_frame, current_map_id,
 		local v = getFilteredClassEntry(_stats["all"][map_id][class_id], selected_source_kind)
 		if v == nil or total_map_entry == nil then
 			entry_data[class_id] = {}
-			local class_str, _, _ = GetClassInfo(class_id)
+			local class_str = class_id == "all" and "all" or select(1, GetClassInfo(class_id))
 			entry_data[class_id]["Class"] = class_str
 			entry_data[class_id]["#"] = "-"
 			entry_data[class_id]["%"] = "-"
@@ -176,7 +176,7 @@ function average_class_container.updateMenuElement(scroll_frame, current_map_id,
 		local total_creature_entry = getCreatureClassEntry(_stats["all"]["all"]["all"], creature_id, selected_source_kind)
 		if v == nil or total_creature_entry == nil then
 			entry_data[class_id] = {}
-			local class_str, _, _ = GetClassInfo(class_id)
+			local class_str = class_id == "all" and "all" or select(1, GetClassInfo(class_id))
 			entry_data[class_id]["Class"] = class_str
 			entry_data[class_id]["#"] = "-"
 			entry_data[class_id]["%"] = "-"