+++ Data/CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 
-## 1774134000
+## 1776290400
 
 ### vanilla
-- 1,333,249 entries as of April 15, 2026
+- 1,333,249 entries as of April 16, 2026
 
 ### tbc
-- 1,335,106 entries as of April 15, 2026
+- 1,335,106 entries as of April 16, 2026